breaking change: use unique resource reservation ids
There are several situations in which slaveIds can change; by using a
UUID for a set of resources, we can match a recovered reservation after
a slave id change to the new slave, and preserve the state.

We can't just use jobId to track reservations, because reservations may
not succeed, creating a situation in which potentially multiple offers
for the same job may exist in the cluster.